Fixed printing error on derived pipeliens

+++ DeduplicateStep.cpp
@@ -15,7 +15,7 @@ vector<string> DeduplicateStep::apply(vector<string> temp){
     string check = temp[0];
     fin.push_back(check);
 
-    for (int i = 1; i < temp.size(); i++)
+    for (int i = 1; i < static_cast<int>(temp.size()); i++)
     {
         if (check != temp[i])
         {
